+++ classes/Generator/Generator/Type/Message.php
@@ -12,4 +12,9 @@ class Generator_Generator_Type_Message extends Generator_Type_Config
 {
 	protected $_folder = 'messages';
 
+	protected function _import_source($source, $path = NULL)
+	{
+		return Generator::get_message($source, $path);
+	}
+
 } // End Generator_Generator_Type_Message