+++ tests/Generator/Type/MessageTest.php
@@ -19,8 +19,6 @@ class Generator_Type_MessageTest extends Unittest_TestCase
 	 */
 	public function test_type_options()
 	{
-		$ds = DIRECTORY_SEPARATOR;
-
 		$type = new Generator_Type_Message();
 		$this->assertSame('messages', $type->folder());
 
@@ -40,6 +38,30 @@ class Generator_Type_MessageTest extends Unittest_TestCase
 			),
 			$params['values']
 		);
+
+		// With imported values
+		$expected = array('two' => array(
+			'three' => 'second message', 'four' => array('five' => 'third message')
+		));
+
+		$type = new Generator_Type_Message();
+		$type->import('testmsgs/generator|two');
+
+		$type->render();
+		$params = $type->params();
+		$this->assertSame($expected, $params['imports']);
+		$this->assertSame($expected, $params['values']);
+
+		// Stored values should override imported values
+		$type = new Generator_Type_Message();
+		$type->import('testmsgs/generator|two');
+		$type->value('two.three|new value');
+
+		$type->render();
+		$params = $type->params();
+		$this->assertSame($expected, $params['imports']);
+		$expected['two']['three'] = 'new value';
+		$this->assertSame($expected, $params['values']);
 	}
 
 } // End Generator_Type_MessageTest